WIP: Fix size mismatch errors occuring during code generation phase
- 'Size mismatch. Input 1 expected value of size [24 x 24]. Found size [:? x :?].
Mismatched varying and fixed sizes indicate a probable run-time error. If this diagnostic is incorrect, use indexing to explicitly make the varying size fixed.': Issue occurs due to imput `H` of `simFunc_qpOASES`. Simulink functions inputs are defined and set through ports, and ports need to have fixed sizes, in this case [3*<num_vertices>,3*<num_vertices>]=[24,24]. For the ports size, <num_vertices> is defined as the constant `contact_config.total_num_vertices`, while `H` size depends on `J_feet`, which is computed by `compute_J_and_JDot_nu()` from the non constant property `num_vertices` of the `Contacts` object. Set `num_vertices` `Contacts` property as `Constant`.
- 'Size mismatch. Input 3 expected value of size [48 x 24]. Found size [:? x 24].
Mismatched varying and fixed sizes indicate a probable run-time error. If this diagnostic is incorrect, use indexing to explicitly make the varying size fixed.': occurs due to wrong size of 'A' built through `blkdiag`.
- 'Function call resolved to MEX file '/Users/nunoguedelha/dev/osqp-matlab/osqp_mex.mexmaci64. Only MATLAB files are supported for code generation': conditional code is compiled even if the respective run condition is false (`obj.useOSQP=false`). So, flags `obj.useOSQP` and `obj.useQPOASES` have to be defined as constants.
- 'Size mismatch (size [0 x 0] ~= size [3 x 29]). The size to the left is the size of the left-hand side of the assignment'. Can't assign a table (e.g. result of an operation) to a table of different size. For that reason, the pattern appending
chunks to a table does not work: A = []; A = [A;B]. Refactor the computation of J, the stacked jacobians relative to the vertices that will be in contact.